Make extra guests conditional a ternary
Before it was using &&. What would happen is that if the number of extra guests was actually 0 (and not null or undefined), a 0 would appear where the cost item would be. Expected behavior is that there's nothing there if there's no extra guests.

Using a ternary to either return the extra guests item or null resolves the issue.